custom_components/elevenlabs-tts/requirements.txt
elevenlabs>=0.2.24
redis>=4.5.0
pyaudio>=0.2.13
numpy>=1.20.0
python-dotenv>=1.0.0
flask>=2.0.0
